Fix Constant types, remove reshapes, rework Gather
Operators no longer manipulate the shape of the output tensors, however, Gather no longer works with the new axis values which are pulled from the node. The tests also prove that ONNX Gather fails to operate on the input data if data is of rank 2, index is of rank 1 and axes=0. Before this change axis was set to a fixed -1.